#1325 - you can open the create form now
