+++ src/frontend/src/pages/ProjectsPage/Projects.tsx
@@ -8,6 +8,7 @@ import { routes } from '../../utils/routes';
 import WBSDetails from '../WBSDetails';
 import CreateWorkPackageForm from '../CreateWorkPackagePage/CreateWorkPackageForm';
 import ProjectsPage from './ProjectsPage';
+import ProjectCreateContainer from '../ProjectDetailPage/ProjectForm/ProjectCreateContainer';
 
 const Projects: React.FC = () => {
   return (
@@ -15,6 +16,7 @@ const Projects: React.FC = () => {
       <Route path={routes.PROJECTS_OVERVIEW} component={ProjectsPage} />
       <Route path={routes.PROJECTS_ALL} component={ProjectsPage} />
       <Route path={routes.WORK_PACKAGE_NEW} component={CreateWorkPackageForm} />
+      <Route path={routes.PROJECTS_NEW} component={ProjectCreateContainer} />
       <Route path={routes.PROJECTS_BY_WBS} component={WBSDetails} />
       <Route path={routes.PROJECTS} component={ProjectsPage} />
     </Switch>
